# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  internal_features_auth_transport_http.CreateUserRequest:
    properties:
      email:
        example: ivanivanov@example.ex
        maxLength: 255
        minLength: 5
        type: string
      full_name:
        example: Ivan Ivanov
        maxLength: 100
        minLength: 3
        type: string
      password:
        example: password_example
        maxLength: 100
        minLength: 6
        type: string
      phone_number:
        example: "+79998887766"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  internal_features_auth_transport_http.CreateUserResponse:
    properties:
      email:
        example: ivanivanov@example.ex
        type: string
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+79998887766"
        type: string
      version:
        example: 3
        type: integer
    type: object
...
  internal_features_auth_transport_http.LogoutUserRequest:
    properties:
      refresh_token:
        example: refresh_token
        maxLength: 255
        minLength: 1
        type: string
    required:
    - refresh_token
    type: object
  internal_features_auth_transport_http.RefreshRequest:
    properties:
      refresh_token:
        example: refresh_token
        maxLength: 255
        minLength: 1
        type: string
    required:
    - refresh_token
    type: object
  internal_features_auth_transport_http.RefreshResponse:
    properties:
      access_token:
        example: access_token
        type: string
    type: object
...
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      description:
        example: Do homework
        maxLength: 1000
        minLength: 1
        type: string
      title:
        example: Homework
        maxLength: 100
        minLength: 1
        type: string
    required:
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Do homework
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Homework
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Do homework
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Homework
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        type: boolean
      description:
        example: "null"
        type: string
      title:
        example: Play basketball
        type: string
    type: object
  internal_features_tasks_transport_http.PatchUserResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Do homework
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Homework
        type: string
      version:
        example: 3
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        example: 5
        type: integer
      completed:
        example: false
        type: boolean
      completed_at:
        example: "null"
        type: string
      created_at:
        example: "2026-02-26T10:30:00Z"
        type: string
      description:
        example: Do homework
        type: string
      id:
        example: 15
        type: integer
      title:
        example: Homework
        type: string
      version:
        example: 3
        type: integer
    type: object
...
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application REST-API scheme
  title: Golang Todo API
  version: "1.0"
paths:
...
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke the provided refresh token for the authenticated user
      parameters:
      - description: LogoutUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.LogoutUserRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Successful user logout
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Refresh token not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Obtain a new access token
      parameters:
      - description: RefreshToken request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Seccessfull refresh token
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.RefreshResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create new user in system
      parameters:
      - description: CreateUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_auth_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Seccessfull created user
          schema:
            $ref: '#/definitions/internal_features_auth_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      summary: Create user
      tags:
      - auth
...
  /tasks:
    get:
      description: View tasks list with optional pagination
      parameters:
      - description: Filter by task id
        in: query
        name: user_id
        type: integer
      - description: Tasks page size
        in: query
        name: limit
        type: integer
      - description: Tasks page shifting
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Seccessfull get a list of tasks
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tasks list
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create new task in system
      parameters:
      - description: CreateTask request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Seccessfull created task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            type: obj
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete an existing task by id
      parameters:
      - description: ID of the task to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successful task deleting
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a specific task by ID
      parameters:
      - description: ID of the received task
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task seccessfuly found
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: |-
        Change existing task information
        ### Logic update fields (Three-state logic):
        1. **The field is not transmitted**: `description` ignored, the value in the database does not change
        2. **Passed value**: `"description`": "Play basketball at 6 pm"` - set new description value
        3. **Passed null**: `"description`": "null"` - clear a field in the database (set to NULL)
        Restrictions: `title ` and `completed` can't be null
      parameters:
      - description: Task id to change
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Seccessfull changed task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_daf32_golang-todoapp_internal_core_transport_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change task
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {access_token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"